# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DZ Bus Tracker API",
    "description": "Real-time bus tracking for Algeria.\n\n* JSON uses snake_case; timestamps are RFC 3339 UTC.\n* Errors are RFC 9457 problem documents (`application/problem+json`) with a stable `code`, localized in French (default), Arabic or English via `Accept-Language`.\n* Lists use cursor pagination: `?limit=` (1–100) and `?cursor=` from `next_cursor`.\n* Cacheable GETs return a weak `ETag`; send `If-None-Match` to get `304`.\n* Rate limits are reported in `RateLimit-*` headers; `429` carries `Retry-After`.",
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/lines/{id}/schedules": {
      "get": {
        "tags": [
          "schedules"
        ],
        "summary": "The schedules of a line by day and start time (not paginated). Active schedules only,\nunless the caller administers schedules.",
        "operationId": "list_line_schedules",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Line id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The schedules of the line",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemList_ScheduleDto"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (If-None-Match)"
          },
          "404": {
            "description": "No such (visible) line",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "schedules"
        ],
        "summary": "Add a schedule to a line (audited). This is the only way to create a schedule.",
        "operationId": "create_schedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Line id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Schedule created (`Location` points to it)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "Missing `schedule:write`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "No such line",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "409": {
            "description": "Overlaps an active schedule of the line on that day (`schedule_overlap`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Invalid fields (e.g. `end_time` not after `start_time`: `must_be_after`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/api/v1/schedules/{id}": {
      "get": {
        "tags": [
          "schedules"
        ],
        "summary": "One schedule. Inactive schedules, and schedules of inactive lines, are only visible to\nschedule administrators.",
        "operationId": "get_schedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Schedule id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The schedule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDto"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (If-None-Match)"
          },
          "404": {
            "description": "No such (visible) schedule",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "schedules"
        ],
        "summary": "Delete a schedule (audited).",
        "operationId": "delete_schedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Schedule id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "Missing `schedule:write`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "No such schedule",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      },
      "patch": {
        "tags": [
          "schedules"
        ],
        "summary": "Update a schedule partially (audited), with the same validation and overlap rules as\ncreation.",
        "operationId": "update_schedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Schedule id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated schedule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "Missing `schedule:write`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "No such schedule",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "409": {
            "description": "Would overlap an active schedule (`schedule_overlap`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Invalid fields",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateScheduleRequest": {
        "type": "object",
        "description": "A new schedule.",
        "required": [
          "day_of_week",
          "start_time",
          "end_time",
          "frequency_minutes"
        ],
        "properties": {
          "day_of_week": {
            "type": "integer",
            "format": "int64",
            "description": "ISO 8601: 1 = Monday … 7 = Sunday.",
            "example": 1
          },
          "end_time": {
            "type": "string",
            "description": "`HH:MM`, after `start_time`.",
            "example": "21:30"
          },
          "frequency_minutes": {
            "type": "integer",
            "format": "int64",
            "description": "1–1440 minutes.",
            "example": 15
          },
          "is_active": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Default `true`."
          },
          "start_time": {
            "type": "string",
            "description": "`HH:MM` (24-hour clock).",
            "example": "06:00"
          }
        },
        "additionalProperties": false
      },
...
      "ItemList_ScheduleDto": {
        "type": "object",
        "description": "A bounded list that is not paginated (e.g. the stops of a line).",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "A weekly service window of a line.",
              "required": [
                "id",
                "line_id",
                "day_of_week",
                "start_time",
                "end_time",
                "frequency_minutes",
                "is_active",
                "created_at",
                "updated_at"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "day_of_week": {
                  "type": "integer",
                  "format": "int32",
                  "description": "ISO 8601: 1 = Monday … 7 = Sunday.",
                  "example": 1,
                  "minimum": 0
                },
                "end_time": {
                  "type": "string",
                  "description": "`HH:MM`, after `start_time`.",
                  "example": "21:30"
                },
                "frequency_minutes": {
                  "type": "integer",
                  "format": "int32",
                  "description": "Minutes between departures (1–1440).",
                  "minimum": 0
                },
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "is_active": {
                  "type": "boolean",
                  "description": "Only active schedules are public; they may not overlap on the same day."
                },
                "line_id": {
                  "type": "string",
                  "format": "uuid"
                },
                "start_time": {
                  "type": "string",
                  "description": "`HH:MM`.",
                  "example": "06:00"
                },
                "updated_at": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          }
        }
      },
...
      "ProblemDocument": {
        "type": "object",
        "description": "RFC 9457 problem document returned by every error response.",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code."
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemFieldError"
            }
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "description": "The request path that produced the problem."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "URI identifying the problem type (`urn:dzbus:problem:<code>`)."
          }
        }
      },
      "ProblemFieldError": {
        "type": "object",
        "description": "A field error in a problem document.",
        "required": [
          "field",
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine code (`required`, `invalid_email`, `too_long`, ...)."
          },
          "field": {
            "type": "string",
            "description": "Path of the offending field (`email`, `scopes[2]`)."
          },
          "message": {
            "type": "string",
            "description": "Localized human-readable message."
          },
          "params": {
            "type": [
              "object",
              "null"
            ],
            "description": "Parameters of the rule (e.g. `{\"max\": 150}`)."
          }
        }
      },
...
      "ScheduleDto": {
        "type": "object",
        "description": "A weekly service window of a line.",
        "required": [
          "id",
          "line_id",
          "day_of_week",
          "start_time",
          "end_time",
          "frequency_minutes",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "day_of_week": {
            "type": "integer",
            "format": "int32",
            "description": "ISO 8601: 1 = Monday … 7 = Sunday.",
            "example": 1,
            "minimum": 0
          },
          "end_time": {
            "type": "string",
            "description": "`HH:MM`, after `start_time`.",
            "example": "21:30"
          },
          "frequency_minutes": {
            "type": "integer",
            "format": "int32",
            "description": "Minutes between departures (1–1440).",
            "minimum": 0
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_active": {
            "type": "boolean",
            "description": "Only active schedules are public; they may not overlap on the same day."
          },
          "line_id": {
            "type": "string",
            "format": "uuid"
          },
          "start_time": {
            "type": "string",
            "description": "`HH:MM`.",
            "example": "06:00"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdateScheduleRequest": {
        "type": "object",
        "description": "Changes to a schedule (same rules as creation).",
        "properties": {
          "day_of_week": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "end_time": {
            "type": [
              "string",
              "null"
            ]
          },
          "frequency_minutes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "is_active": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "start_time": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Machine-to-machine key (`dzk_…`) with explicit scopes."
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from /auth/login or /auth/refresh (EdDSA)."
      }
    }
  },
...
}